fix: Se configura y realiza la sección de reservas, se está a la espera de confirmación.
