+++ src/js/urlData.js
@@ -1,0 +1,9 @@
+export function getUrlSpaces() {
+    const URLSpaces = `http://localhost:3000/spaces`;
+    return URLSpaces;
+}
+
+export function getUrlUsers() {
+    const URLUsers = `http://localhost:3000/users`;
+    return URLUsers;
+}
